여기땅

+++ PuppyRun/src/community/model/controller/CommunityDeleteServlet.java
@@ -3,6 +3,7 @@ package community.model.controller;
 import java.io.File;
 import java.io.IOException;
 import java.io.PrintWriter;
+import java.util.ArrayList;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -12,6 +13,7 @@ import javax.servlet.http.HttpServletResponse;
 
 import community.model.service.CommentService;
 import community.model.service.CommunityService;
+import community.model.vo.Comment;
 import community.model.vo.Community;
 import photo.model.service.PhotoService;
 
@@ -28,19 +30,29 @@ public class CommunityDeleteServlet extends HttpServlet {
 		Community community = new CommunityService( ).selectOneCommunity(communityNo);
 		String communityPhoto = community.getComPhoto();
 		String communityUserId = community.getComId();
-		int commentResult = 0;
 		int result = 0;
 		
-		commentResult = new CommentService().deleteCommunity2(communityNo);
-		result = new CommunityService().deleteCommunity(communityNo);
+		ArrayList<Comment> cList = new CommentService().selectAllComment(communityNo);
+		int commentResult = 0;
+		if(!cList.isEmpty()) {
+			commentResult = new CommentService().deleteCommunity2(communityNo);
+		} else {
+			commentResult = 1;
+		}
 		
-		int photoResult = 1;
+		// 파일 삭제
+		int photoResult = 0;
 		if(communityPhoto != null) {
 			String photoPath = new PhotoService().selectPhoto(communityPhoto, communityUserId);
 			photoResult = new PhotoService().removePhoto(communityPhoto, communityUserId);
 			File file = new File(photoPath);
 			file.delete(); //위에 가져온 파일 경로와 동일한 파일 삭제
+		} else {
+			photoResult = 1;
 		}
+		
+		// 게시글 삭제
+		result = new CommunityService().deleteCommunity(communityNo);
 		
 		if(result > 0 && photoResult > 0 && commentResult > 0) {
 			response.setContentType("text/html; charset=UTF-8");
